+++ app/Http/Livewire/Series/EditSeries.php
@@ -2,7 +2,6 @@
 
 namespace App\Http\Livewire\Series;
 
-use App\Http\Traits\LivewireDelete;
 use App\Models\Category;
 use App\Models\Publisher;
 use App\Models\Series;
@@ -67,7 +66,7 @@ class EditSeries extends Component
         return view('livewire.series.edit-series')->extends('layouts.app')->section('content');
     }
 
-    public function save(): void
+    public function save()
     {
         $this->validate();
         if (!empty($this->series->default_price)) {
@@ -78,8 +77,10 @@ class EditSeries extends Component
             $this->series->save();
             $this->storeImages($image);
             $this->updatePrices();
-            toastr()->livewire()->addSuccess(__(':name has been updated', ['name' => $this->series->name]));
-            $this->reset(['image_url']);
+            $this->updateStatuses();
+            toastr()->addSuccess(__(':name has been updated', ['name' => $this->series->name]));
+
+            return redirect()->route('series.show', [$this->category, $this->series]);
         } catch (Exception $exception) {
             Log::error($exception);
             toastr()->livewire()->addError(__(':name could not be updated', ['name' => $this->series->name]));
@@ -129,6 +130,13 @@ class EditSeries extends Component
     {
         if (!empty($this->series->default_price) && $this->series->default_price > 0) {
             Volume::whereSeriesId($this->series->id)->whereNull('price')->update(['price' => $this->series->default_price]);
+        }
+    }
+
+    private function updateStatuses(): void
+    {
+        if ($this->series->subscription_active) {
+            Volume::whereSeriesId($this->series->id)->where('status', '=', '0')->update(['status' => 1]);
         }
     }
 }