+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 2.1.0 (2022-02-07)
+
+- Added date format to user profile
+- Improve updating series and volumes data Manga Passion API
+- ISBN and Publish Date of a volume are now optional
+- Reordering volumes makes sure that no gaps are left
+- When setting a series as subscribed, all volumes with status "New" will be set to "Ordered"
+
 ## 2.0.3 (2022-02-05)
 
 - Applied fixes from 2.0.2 to update button